add evaluation status
